Avoid recreating timelapse MP4 if already exists
Disables time and space consuming timalapse generation, if exists from previous run.
If needed, file can be manually removed and timelapse will be created again.

+++ iStream/iStream.sh
@@ -94,16 +94,18 @@ function generate_captured_stack {
 }
 
 function generate_timelapse {
-	SECONDS_LIMIT=$(expr $REMAINING_SECONDS - 120)
-	if ($LEVEL_1); then
-		cd ~/source/RMS
-		python -m Utils.GenerateTimelapse $CAPTURED_DIR_NAME
-		mv $TMP_VIDEO_FILE $VIDEO_FILE
-	fi
-	if ($LEVEL_2); then
-		cd ~/source/RMS
-		python -m Utils.GenerateTimelapse $CAPTURED_DIR_NAME
-		mv $TMP_VIDEO_FILE $VIDEO_FILE
+	if [ ! -f "$VIDEO_FILE" ]; then
+		SECONDS_LIMIT=$(expr $REMAINING_SECONDS - 120)
+		if ($LEVEL_1); then
+			cd ~/source/RMS
+			python -m Utils.GenerateTimelapse $CAPTURED_DIR_NAME
+			mv $TMP_VIDEO_FILE $VIDEO_FILE
+		fi
+		if ($LEVEL_2); then
+			cd ~/source/RMS
+			python -m Utils.GenerateTimelapse $CAPTURED_DIR_NAME
+			mv $TMP_VIDEO_FILE $VIDEO_FILE
+		fi
 	fi
 }
